Abstract - Created move_maker class prototype. - Heavily refactored board.py - Added xor and not to bitboard.py - Moved CastlingRights to its own file. - Corrected incorrect dunder methods and type hints in board.py - Added get_piece method to board.py
Structure Changes
The Board class is no longer a frozen dataclass.

Concerns
The MoveMaker class was fairly annoying to create because of the way pieces are encoded. It also appears to be very difficult to write proper tests for. Both of these things will need to be fixed in the future.